requirements.txt
langchain
streamlit 
dotenv
langchain-core
langchain-community
sse-starlette
pypdf
faiss-cpu
pymupdf
langchain-google-genai
uvicorn
fastapi
tiktoken
python-dotenv
requests
python-multipart
shutil
hashlib
google-generativeai
pathlib